# Conflicts:
#	.travis.yml
#	gemrb/GUIScripts/GUIMG.py
#	gemrb/GUIScripts/GUISTORE.py
#	gemrb/GUIScripts/LUSkillsSelection.py
#	gemrb/GUIScripts/iwd2/Appearance.py
#	gemrb/core/CharAnimations.cpp
#	gemrb/core/ControlAnimation.cpp
#	gemrb/core/GUI/Button.cpp
#	gemrb/core/GUI/GameControl.cpp
#	gemrb/core/GUI/GameControl.h
#	gemrb/core/GUI/MapControl.h
#	gemrb/core/GameScript/GSUtils.h
#	gemrb/core/Holder.h
#	gemrb/core/Interface.cpp
#	gemrb/core/Map.cpp
#	gemrb/core/Map.h
#	gemrb/core/Scriptable/Actor.cpp
#	gemrb/core/Scriptable/Actor.h
#	gemrb/core/Scriptable/Door.h
#	gemrb/core/Scriptable/Scriptable.cpp
#	gemrb/core/ScriptedAnimation.h
#	gemrb/core/TileMap.cpp
#	gemrb/plugins/AREImporter/AREImporter.cpp
#	gemrb/plugins/AREImporter/AREImporter.h
#	gemrb/plugins/CHUImporter/CHUImporter.cpp
#	gemrb/plugins/FXOpcodes/FXOpcodes.cpp
#	gemrb/plugins/GUIScript/GUIScript.cpp
#	gemrb/plugins/IWDOpcodes/IWDOpcodes.cpp
#	gemrb/plugins/TTFImporter/TTFFont.cpp

+++ gemrb/plugins/BIKPlayer/fft.cpp
@@ -117,8 +117,8 @@ av_cold int ff_fft_init(FFTContext *s, int nbits, int inverse)
 
         for(i=0; i<(n/2); i++) {
             alpha = (float) (2 * M_PI * (float)i / (float)n);
-            c1 = (float) cos(alpha);
-            s1 = (float) (sin(alpha) * s2);
+            c1 = cos(alpha);
+            s1 = sin(alpha) * s2;
             s->exptab[i].re = c1;
             s->exptab[i].im = s1;
         }
